+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 * Ozon Content: Renvooi op `tekst:row` ([#3382](https://github.com/dso-toolkit/dso-toolkit/issues/3382))
+* Ozon Content: Begrip, Term en Definitie incl. renvooi ([#3412](https://github.com/dso-toolkit/dso-toolkit/issues/3412))
 
 ### Fixed
 * Ozon Content: `wijzigactie` van `opschrift` in `kop` toont geen renvooi ([#3413](https://github.com/dso-toolkit/dso-toolkit/issues/3413))